+++ objects
@@ -1,5 +1,17 @@
-4
-sph 10 2 1 1 1.0 1.0 0.0 1000
-box 3 4 5 1 1 3 0.0 0.0 1.0 1
-box 5 5 10 1 1 1 0.0 1.0 1.0 0
-pla 0 0 1 0.1 0.7 0.1 0.01
+10
+pla
+0.0 0.0 1.0
+0.1 0.7 0.1 0.5
+sph
+10.0 0.0 11.0 1.3
+1.0 0.0 0.0 0.0
+sph
+10.0 -2.0 6.0 1.5
+0.0 1.0 0.0 1000.0
+sph
+10.0 2.0 6.0 1.5
+0.0 1.0 0.0 1000.0
+box
+10.0 0.0 8.0
+1.0 1.0 3.0
+1.0 0.0 0.0 0.0
